# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skool.loc",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1/"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Post a new data",
                "description": "Register",
                "operationId": "da5e12b664c1b1601a83bbcde0ddbf1f",
                "requestBody": {
                    "description": "data  credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "email",
                                    "password",
                                    "user_bio",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Salohiddin"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Nuriddinov"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nuridinovsaloxiddin@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "admin123"
                                    },
                                    "user_bio": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "User haqida malumot"
                                    },
                                    "address": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "Address: Samarqand"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/activate-user": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verification",
                "description": "Ro'yxatdan o'tgandan so'ng userni aktiv qilish",
                "operationId": "faca470a27aa9665997b78169f849fb1",
                "requestBody": {
                    "description": "Verify",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "verify_code"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nuridinovsaloxiddin@gmail.com"
                                    },
                                    "verify_code": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": 123456
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "SUCCESS"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "login",
                "description": "Login emaill, password",
                "operationId": "00179ad2a04cd490eabed0a6138eac05",
                "requestBody": {
                    "description": "Pass credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "user@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/revoke-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "password revoke",
                "description": "revoke password",
                "operationId": "4e1536c98f648ffa640ffeb6f7a4d4af",
                "requestBody": {
                    "description": "Pass credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "user@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/reset-password-verify": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "reset password",
                "description": "Parolni tiklash uchun tasdiqlash kodi",
                "operationId": "29ecf88d489e59b838c8209c50332fbf",
                "requestBody": {
                    "description": "Verify",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "verify_code",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "nuridinovsaloxiddin@gmail.com"
                                    },
                                    "verify_code": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": 123456
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "user12345"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "user12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "SUCCESS"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "title": "ResponseError",
                "properties": {
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "text",
                            "example": "Not Found"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "admin": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}